rename elf.rs test to loader.rs; move pe loading test into loader. add mach dylib and exe, along with pe dll to tests
